backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
httpx>=0.26.0
numpy>=1.24.0
python-multipart>=0.0.6
pymupdf>=1.24.0
aiofiles>=23.2.0
easyocr>=1.7.0
Pillow>=10.0.0
torch>=2.0.0
